requirements.txt
streamlit==1.34.0
youtube-transcript-api==0.6.2
python-docx==1.1.0
